Class separation
Separated classes into different files, set up test files similarly, added placeholder tests for various methods.
