+++ karenina/__init__.py
@@ -1,1 +1,2 @@
-#Blank init file 
+#Blank init file
+# TODO: setup init to run module
