+++ log.txt
@@ -1,0 +1,11 @@
+#Karenina Simulation Logfile
+Output folder: C:/Users/Ripti/Dropbox/Peoples/CSS143/karenina-master
+Treatment names: control,destabilizing_treatment
+Number of individuals: 35,35
+Number of timepoints: 10
+Perturbation timepoint: 5
+Perturbation duration: 100
+Interindividual variation: 0.01
+Delta: 0.25
+Lambda: 0.2
+Fixed starting position: None
