+++ tests/test_spatial_ornstein_uhlenbeck_individual.py
@@ -9,30 +9,42 @@ __maintainer__ = "Jesse Zaneveld"
 __email__ = "zaneveld@gmail.com"
 __status__ = "Development"
 
-import unittest 
+import unittest
 from warnings import catch_warnings
-from karenina.spatial_ornstein_uhlenbeck  import Process,Individual,Experiment,Perturbation
+from karenina.spatial_ornstein_uhlenbeck import Process, Individual, Experiment, Perturbation
 import numpy.testing as npt
-  
+
 """
 Tests for spatial_ornstein_uhlenbeck.py
 """
 
-class TestPrimary(unittest.TestCase):
+
+class TestIndividual(unittest.TestCase):
     # TODO: Tests
 
     def setUp(self):
         pass
 
-    def test_check_perturbation_timepoint(self):
+    def test_applyPerturbation(self):
         pass
 
-    def test_write_options_to_log(self):
+    def test_removePerturbation(self):
         pass
 
-    def test_parse_perturbation_file(self):
+    def test_removePerturbationFromAxis(self):
         pass
-    
+
+    def test_applyPerturbationToAxis(self):
+        pass
+
+    def test_check_identity(self):
+        pass
+
+    def test_simulate_movement(self):
+        pass
+
+    def test_get_data(self):
+        pass
 
 if __name__ == '__main__':
     unittest.main()
